Moving the grid and player into their own separate files
