+++ src/components/collection/container.styl
@@ -66,7 +66,7 @@
   width: 85px
   height: 75px
 
-.playControlContainer
+.playOrGridToggle
   background: overlay-background
   padding: 12px 22px
